refactor: centralize actix-http test service
