+++ actix-http/src/test/mod.rs
@@ -3,6 +3,8 @@
 mod test_buffer;
 mod test_request;
 mod test_seq_buffer;
+#[cfg(test)]
+pub(crate) mod test_services;
 
 pub use self::{
     test_buffer::TestBuffer,
